# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeClash - Real-time competitive coding battle platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List public battle rooms
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [waiting, active, finished]
        - name: battleType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"

    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new battle room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{roomCode}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room by code
      parameters:
        - name: roomCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Not found
  /rooms/{roomCode}/join:
    post:
      operationId: joinRoom
      tags: [rooms]
      summary: Request to join a room
      parameters:
        - name: roomCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRoomRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
...
  /rooms/{roomCode}/start:
    post:
      operationId: startBattle
      tags: [rooms]
      summary: Start the battle (host only)
      parameters:
        - name: roomCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Battle"
  /rooms/{roomCode}/approve/{userId}:
    post:
      operationId: approveJoinRequest
      tags: [rooms]
      summary: Approve/reject join request (host only)
      parameters:
        - name: roomCode
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinApproval"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

  # ── Battles ──
...
components:
  schemas:
...
    Room:
      type: object
      required: [id, roomCode, hostId, hostUsername, battleType, difficulty, status, maxPlayers, requireApproval, isPrivate, participants, createdAt]
      properties:
        id:
          type: integer
        roomCode:
          type: string
        hostId:
          type: integer
        hostUsername:
          type: string
        battleType:
          type: string
          enum: [duel, deathmatch, multiplayer, topic_battle, battle_royale, arena]
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
        topic:
          type: ["string", "null"]
        status:
          type: string
          enum: [waiting, starting, active, finished]
        maxPlayers:
          type: integer
        requireApproval:
          type: boolean
        isPrivate:
          type: boolean
        inviteLink:
          type: ["string", "null"]
        participants:
          type: array
          items:
            $ref: "#/components/schemas/RoomParticipant"
        battleId:
          type: ["integer", "null"]
        createdAt:
          type: string
    RoomInput:
      type: object
      required: [battleType, difficulty, maxPlayers]
      properties:
        battleType:
          type: string
          enum: [duel, deathmatch, multiplayer, topic_battle, battle_royale, arena]
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
        topic:
          type: string
        maxPlayers:
          type: integer
        requireApproval:
          type: boolean
        isPrivate:
          type: boolean
    RoomParticipant:
      type: object
      required: [userId, username, cfHandle, rating, rank, isReady, isHost]
      properties:
        userId:
          type: integer
        username:
          type: string
        cfHandle:
          type: string
        rating:
          type: integer
        rank:
          type: string
        isReady:
          type: boolean
        isHost:
          type: boolean
        joinedAt:
          type: string
    JoinRoomRequest:
      type: object
      properties:
        roomCode:
          type: string
    JoinApproval:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean

    # ── Battles ──
    Battle:
      type: object
      required: [id, roomId, roomCode, battleType, status, problem, participants, startedAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        roomCode:
          type: string
        battleType:
          type: string
        status:
          type: string
          enum: [active, finished]
        problem:
          $ref: "#/components/schemas/Problem"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/BattleParticipantStatus"
        startedAt:
          type: string
        endedAt:
          type: ["string", "null"]
        winnerId:
          type: ["integer", "null"]
        winnerUsername:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
    BattleParticipantStatus:
      type: object
      required: [userId, username, cfHandle, rating, rank, solved, attempts, penalty, rank_position]
      properties:
        userId:
          type: integer
        username:
          type: string
        cfHandle:
          type: string
        rating:
          type: integer
        rank:
          type: string
        solved:
          type: boolean
        solveTimeSeconds:
          type: ["integer", "null"]
        attempts:
          type: integer
        penalty:
          type: integer
        rank_position:
          type: integer
        isEliminated:
          type: boolean
        lastVerdict:
          type: ["string", "null"]
...
    Problem:
      type: object
      required: [id, contestId, problemIndex, title, rating, tags, cfUrl]
      properties:
        id:
          type: integer
        contestId:
          type: integer
        problemIndex:
          type: string
        title:
          type: string
        rating:
          type: integer
        tags:
          type: array
          items:
            type: string
        cfUrl:
          type: string
        solvedCount:
          type: ["integer", "null"]
...