# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeClash - Real-time competitive coding battle platform
 * OpenAPI spec version: 0.1.0
 */
import type { RoomBattleType } from './roomBattleType';
import type { RoomDifficulty } from './roomDifficulty';
import type { RoomParticipant } from './roomParticipant';
import type { RoomStatus } from './roomStatus';

export interface Room {
  id: number;
  roomCode: string;
  hostId: number;
  hostUsername: string;
  battleType: RoomBattleType;
  difficulty: RoomDifficulty;
  /** @nullable */
  topic?: string | null;
  status: RoomStatus;
  maxPlayers: number;
  requireApproval: boolean;
  isPrivate: boolean;
  /** @nullable */
  inviteLink?: string | null;
  participants: RoomParticipant[];
  /** @nullable */
  battleId?: number | null;
  createdAt: string;
}